+++ packages/confirmation/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './countdown';
 export * from './sign-confirmation';
+export * from './countdown-loader';