+++ packages/input/src/component.screenshots.test.tsx
@@ -16,7 +16,7 @@ const screenshotTesting = setupScreenshotTesting({
 const clip = { x: 0, y: 0, width: 350, height: 150 };
 
 describe('Input | screenshots main props', () => {
-    const testCase = (theme: string) =>
+    const testCase = (theme: string, colors: string) =>
         screenshotTesting({
             cases: [
                 [
@@ -25,10 +25,11 @@ describe('Input | screenshots main props', () => {
                         componentName: 'Input',
                         knobs: {
                             label: 'Label',
-                            size: ['s', 'm', 'l'],
+                            size: ['s', 'm', 'l', 'xl'],
                             block: [false, true],
                             disabled: [false, true],
                             value: ['', 'Value'],
+                            colors,
                         },
                         size: { width: 350, height: 150 },
                     }),
@@ -40,11 +41,13 @@ describe('Input | screenshots main props', () => {
             theme,
         })();
 
-    ['default', 'click', 'site', 'mobile'].map(testCase);
+    ['default', 'inverted'].forEach(colors => {
+        ['default', 'click', 'site', 'mobile'].forEach(theme => testCase(theme, colors));
+    });
 });
 
 describe('Input | screenshots hint and error', () => {
-    const testCase = (theme: string) =>
+    const testCase = (theme: string, colors: string) =>
         screenshotTesting({
             cases: [
                 [
@@ -56,6 +59,7 @@ describe('Input | screenshots hint and error', () => {
                             hint: ['', 'Hint'],
                             error: ['', 'Error'],
                             value: ['', 'Value'],
+                            colors,
                         },
                         size: { width: 350, height: 150 },
                     }),
@@ -67,7 +71,36 @@ describe('Input | screenshots hint and error', () => {
             theme,
         })();
 
-    ['default', 'click', 'site', 'mobile'].map(testCase);
+    ['default', 'inverted'].forEach(colors => {
+        ['default', 'click', 'site', 'mobile'].forEach(theme => testCase(theme, colors));
+    });
+});
+
+describe('Input | clear icon', () => {
+    const testCase = (theme: string, colors: string) =>
+        screenshotTesting({
+            cases: [
+                [
+                    `${theme} theme`,
+                    createSpriteStorybookUrl({
+                        componentName: 'Input',
+                        knobs: {
+                            size: 'm',
+                            clear: true,
+                            value: 'value',
+                            colors,
+                        },
+                        size: { width: 350, height: 150 },
+                    }),
+                ],
+            ],
+            screenshotOpts: {
+                fullPage: true,
+            },
+            theme,
+        })();
+
+    ['default', 'inverted'].forEach(colors => testCase('default', colors));
 });
 
 describe(